backend/requirements.txt
# backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mysqlclient>=2.2
boto3>=1.34
django-storages>=1.14
python-magic>=0.4.27
Pillow>=10.2
gunicorn>=21.2
python-decouple>=3.8
